Images can be handled in recipe steps, main view maintains scroll position on rotation, videos stay paused/continue playing on device rotation, widget displays ingredients list

+++ app/src/main/java/com/example/android/bakingtime/adapters/IngredientsAdapter.java
@@ -84,7 +84,7 @@ public class IngredientsAdapter extends RecyclerView.Adapter<IngredientsAdapter.
     private Ingredients getIngredients(int position) {
         mCursor.moveToPosition(position);
 
-        int quantity = mCursor.getInt(DetailFragment.COL_QUANTITY);
+        double quantity = mCursor.getDouble(DetailFragment.COL_QUANTITY);
         String measure = mCursor.getString(DetailFragment.COL_MEASURE);
         String ingredient = mCursor.getString(DetailFragment.COL_INGREDIENT);
